requirements.txt
streamlit
pandas
scikit-learn
Pillow
openpyxl
base64